refactor(similarity): split section CRUD into focused adapter
The new BookSimilarityEmbeddingRepository was mixing two persistence
concerns (book-level fused vectors and per-section embedding cache)
and sat at 449 lines — well above the 350-line ceiling in AGENTS.md
[LOC1a]. Flagged by greptile review on PR #91.

Extract the section-embedding cache into its own focused adapter so
each repository owns one storage concern. The service now depends on
both repositories; nothing outside the service touches the section
cache, so the split is invisible to the rest of the application. Also
collapsed the two type-specific `serialize` helpers into one generic
`<T> serialize(T, String)` method so the remaining file can trim
duplicated exception handling.

- Introduce BookEmbeddingSectionRepository (107 lines) owning
  `fetchSectionEmbedding`, `upsertSectionEmbedding`, `INPUT_PREVIEW`
  truncation, and the `IdGenerator` usage for section rows
- Remove section CRUD, `preview` helper, and `BookSimilaritySection*`
  imports from BookSimilarityEmbeddingRepository
- Replace two overloaded `serialize(SourceMetadata,...)` /
  `serialize(SectionHashPayload,...)` methods with one generic
  `<T> serialize` and remove the `toJson` wrappers
- Inject BookEmbeddingSectionRepository into
  BookSimilarityEmbeddingService and reroute section reads/writes
- BookSimilarityEmbeddingRepository drops from 449 to 367 lines
